% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptFonctionsSourcesVpack.r
\name{anovLetters}
\alias{anovLetters}
\title{AnovLetters}
\usage{
anovLetters(VAR, FAC, ALPHA = 0.05)
}
\arguments{
\item{VAR}{variable}

\item{FAC}{facteur}

\item{ALPHA}{(facultatif) donne le seuil de significativit?, 0.05 par defaut...}
}
\value{
une liste de lettres qui correspondent aux modalites d'un facteur
}
\description{
Donne les lettres associees a un test post hocc de Tukey effectue apres une anova (implique donc de respecter les conditions d'utilisation des tests parametriques...)
}
\examples{
biomasse<-c(rnorm(35,20,5),rnorm(35,15,3),rnorm(35,5,3))
traitement<-factor(rep(c("T1","T2","T3"),each=35))
anov<-aov(biomasse~traitement)
summary(anov)
# OK, il y a un effet significatif, du coup anovLetters renvoie la liste des lettres...
lettresPH<-anovLetters(VAR = biomasse,FAC = traitement)
lettresPH
# ...que l'on peut utiliser directement sur barres.plot :
barres.plot(biomasse,traitement,lettres = lettresPH)
}
