% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptFonctionsSourcesVpack.r
\name{barres.plot}
\alias{barres.plot}
\title{Diagramme en barre}
\usage{
barres.plot(
  variable,
  Facteur,
  lettres = c(""),
  ecart = "sem",
  ylim = "NP",
  ...
)
}
\arguments{
\item{variable}{un vecteur avec les valeurs des variables}

\item{Facteur}{un facteur donnant les differentes modalites}

\item{lettres}{(facultatif) liste de type c("text","text") qui figurera au dessus des barres}

\item{ecart}{(facultatif) type de calcul de la barre d'erreur (sem, IC, sd, var, etc.)}

\item{ylim}{comme pour la fonction plot(), permet de preciser les limites de l'axe des ordonnees}

\item{...}{possibilite de rajouter des arguments comme col,main,ylab,etc. associees a barplot}
}
\description{
Fait un diagramme en barre avec la moyenne et les barres d'erreurs, possibilite de changer le calcul des barres d'erreur, et d'ajouter les lettres des tests post hoc
}
\examples{
#Donnees
biomasse<-c(rnorm(35,20,5),rnorm(35,15,3),rnorm(35,5,3))
traitement<-factor(rep(c("T1","T2","T3"),each=35))
#Graphique de base
barres.plot(biomasse,traitement)
#En changeant les barres d'erreurs
barres.plot(biomasse,traitement,ecart=sd)
#En ajoutant les lettres de post-hoc
barres.plot(biomasse,traitement,lettres=c("a","ab","b"))
#En changeant des parametres classiques de barplot...
barres.plot(biomasse,traitement,lettres=c("a","ab","b"),
col=c("cadetblue","gold","orange"),
xlab="Traitements herbicides",ylab="Biomasse",font.lab=3,
main="Efficacite des traitement")
#Pour Julie.., si on veut changer l'ordre des traitements, il faut reprendre la fonction factor()
traitement<-factor(traitement,levels=c("T3","T1","T2"))
barres.plot(biomasse,traitement)

}
\seealso{
\code{\link{barplot}} pour tout les arguments dans \code{...}
}
