% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptFonctionsSourcesVpack.r
\name{BBtransf}
\alias{BBtransf}
\title{transformation de Braun-Blanquet}
\usage{
BBtransf(comtab, plus = 0.1)
}
\arguments{
\item{comtab}{un tableau de releves de vegetation}

\item{plus}{la valeurs par lequel le + a ete transcrit... souvent c'est 0.1, 0.2 ou 0.5 (par defaut ici c'est 0.1)}
}
\value{
un tableau avec les valeurs transformees
}
\description{
Transforme les coef de Braun-Blanquet en pourcentage (moyenne des classes) Guinochet M. (1973). Phytosociologie. Masson, Paris
}
