% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptFonctionsSourcesVpack.r
\name{better_arrows}
\alias{better_arrows}
\title{Trace des fleches suite a une ACP}
\usage{
better_arrows(COO, coef = 1, length = 0.1, angle = 15, col_text = 1, ...)
}
\arguments{
\item{COO}{coordonnées des variables (ANALYSE$co pour une ACP avec ade4::dudi.pca et summary(ANALYSE)$biplot pour une RDA avec vegan::rda)}

\item{coef}{un coef multiplicateur (pour exploser ou non la dispersion dans le plan)}

\item{length}{longueur de l'encoche des fleches, 0.1 = 10pourcent par defaut}

\item{angle}{angle des encoches des fleches, 15 = 15deg par defaut}

\item{col_text}{couleur du texte}

\item{...}{n'importe quel argument de la fonction arrows}
}
\description{
Trace des fleches suite a une ACP
}
\examples{
#après une rda :
library(vegan)
data("dune")
data("dune.env")
A1<- dune.env$A1
Manure <- as.numeric(as.character(dune.env$Manure))
Moisture <- as.numeric(as.character(dune.env$Moisture))
ANALYSE_RDA <- rda(dune~A1+Manure+Moisture)
plot(summary(ANALYSE_RDA)$sites)
better_arrows(COO = summary(ANALYSE_RDA)$biplot, coef = 2, col_text = "orange")
# après une acp :
library(ade4)
ANALYSE_ACP <- dudi.pca(dune, scannf = F, nf = 2) #pas idéal sur des données espèces, c'est juste pour l'exemple..
plot(ANALYSE_ACP$li) #pour avoir le cadre des relevés
better_arrows(COO = ANALYSE_ACP$co, coef = 3, col_text = "firebrick")
}
