% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptFonctionsSourcesVpack.r
\name{Classes_def}
\alias{Classes_def}
\title{Definition de classes}
\usage{
Classes_def(variab, bornes, noms_classes)
}
\arguments{
\item{variab}{un vecteur a transformer}

\item{bornes}{les bornes definissant les classes, la borne inferieure de la classe la plus basse ainsi que la borne superieure de la classe la plus haute ne doivent pas etre renseignees}

\item{noms_classes}{les noms des classes par lequel sera transforme le vecteur variab. Doit avoir une dimensions de length(bornes)+1}
}
\description{
Permet de transformer un vecteur d'une variable numerique en classes
}
\examples{
variable<-c(0.5, #classe 1
1.2,1.3, #classe 2
2.1, #classe 3
7.2,8.4,9, #classe 5
10.5, #classe 6
22) #et classe 7

boundaries<-c(1,2,3,7,10,15)
classnames<-paste("classe",c(1:(length(bornes)+1)))
Classes<-Classes_def(variable,boundaries,classnames)
data.frame(Classes,variable)
}
