% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptFonctionsSourcesVpack.r
\name{combin.tab}
\alias{combin.tab}
\title{Combinaison de tableau}
\usage{
combin.tab(tableau1, tableau2)
}
\arguments{
\item{tableau1}{le premier tableau, avec les variables en colonnes et les releves en ligne}

\item{tableau2}{le deuxieme tableau, comme le tableau1}
}
\value{
un tableau combine
}
\description{
Permet de combiner 2 tableaux (de contingence ou pas), en fusionnant les colonnes ayant le meme nom
}
\details{
Si on veut combiner plus de 2 tableaux, il vaut mieux faire un \code{\link{write.table}} et \code{\link{read.table}} pour chaque tableau intermediaire (j'ai deja eu quelques bugs...)

Par defaut, les variables seront classees par ordre alphabetique, si vous voulez garder les premieres colonnes (quadrat, date, traitement, hauteur de veg) au debut,
il vaut mieux utiliser l'ancienne version de la fonction \code{\link{combin.tabV0}}, qui demande un autre argument \code{Liste} une liste des noms des colonnes, pour la faire, sous excel, c'est simple, si vous copiez les noms d'especes a la suite (sur une meme colonnes), vous faites supprimer les doublons apres l'avoir mis par ordre alaphabetique, il suffit de rajouter les variables d'information au debut de la liste, et voila, vous avez votre liste en txt!
}
