% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptFonctionsSourcesVpack.r
\name{ComStructIndices}
\alias{ComStructIndices}
\title{Community Structure Integrity Indices}
\usage{
ComStructIndices(REF, ASSESS, rar = 1)
}
\arguments{
\item{REF}{is the reference community data matrix}

\item{ASSESS}{is the assessed community data matrix}

\item{rar}{(facultative) Minimum number of samples in which species have to be present to be taken into account in the calculation of indices. Default value is 1. It should not be used in the indices calculation, but it can be useful to reduce the number of species with the structure.plot() function.}
}
\value{
\item{Comb}{A combined community data matrix of reference and assessed communities}

\item{Nam_Tot}{A list of species names corresponding to the Comb matrix}

\item{Nam_Tar}{A list of the target species names}

\item{REF_Tab}{Reference community data matrix (with zero values for species which were absent in the reference community)}

\item{ASSESS_Tab}{Assessed community data matrix (with zero values for species which were absent in the assessed community)}

\item{SumMeanAbREF}{Sum of mean abundances}

\item{SumAbASSESS}{A list of sum of species abundance for each sample}

\item{Diff}{Matrix of differences, for each species, between mean abundance in the reference samples and abundance in each assessed community samples}

\item{SumNeg}{Sum of negative differences issued from Diff, i.e. sum of 'higher abundances' in each of assessed community sample}

\item{SumPos}{Sum of positive differences issued from Diff, i.e. sum of 'missing abundances' in each of assessed community sample}

\item{CSII}{A list of Community Integrity Index in each assessed community sample}

\item{HAI}{A list of Higher Abundance Index in each of assessed community sample}

\item{CSIInorm}{A list of Normalized Community Integrity Index in each assessed community sample}

\item{AbMeanREFOnly}{A list of mean abundances of target species in reference samples}

\item{ASSESSTarOnly_Tab}{An assessed community data matrix with target species only}

\item{HigherOnly_Tab}{An assessed community data matrix with non-target species only}
}
\description{
Calculates indices of community integrity compared to a reference community according to Jaunatre et al. (2013).
}
\seealso{
\code{\link{structure.plot}} for graphical output
}
