% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptFonctionsSourcesVpack.r
\name{Couleur_continue}
\alias{Couleur_continue}
\title{Couleur continue}
\usage{
Couleur_continue(
  VAR,
  COLORS = c("Red", "Blue"),
  minVAR = "def",
  maxVAR = "def",
  nombre = 100
)
}
\arguments{
\item{VAR}{un vecteur}

\item{COLORS}{une liste de couleur cf l'argument colors de la fonction \code{\link{colorRampPalette}}}

\item{minVAR}{valeur minimale theorique (le minimum de VAR par defaut)}

\item{maxVAR}{valeur maximale theorique (le maximum de VAR par defaut)}

\item{nombre}{precision de la couleur, c-a-d le nombre d'elements composant la palette de couleur (100 par defaut)}
}
\value{
un vecteur de couleur de la meme longueur que VAR
}
\description{
Transforme une variable en palette de couleur
}
\examples{
xx<-c(1:11)
yy<-c(1,3,3.5,4,8,9,3,2.5,7,7.2,NA)
plot(xx,xx,cex=5,col=Couleur_continue(yy),pch=15)
plot(xx,xx,cex=5,col=Couleur_continue(yy,COLORS=c("Orange","Green")),pch=15)
plot(xx,xx,cex=5,col=Couleur_continue(yy,COLORS=c("Orange","Green"),maxVAR=100),pch=15)
}
