% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptFonctionsSourcesVpack.r
\name{Diss_Ref_Plot}
\alias{Diss_Ref_Plot}
\title{Plot of references and restoration site dissimilarities}
\usage{
Diss_Ref_Plot(
  RELEVES,
  REF,
  DISTANCES,
  METHOD = "euclidean",
  COUL_RELEVES = 2,
  COUL_Rel_variable = TRUE,
  COUL_REF = 1,
  COUL_Seg = "#9C8809",
  COUL_Seg_variable = TRUE,
  LINK_NUMBER = "absent",
  N_REF = "absent",
  DIST_MIN = "absent",
  VAL_DIST = TRUE,
  DECAL = 0
)
}
\arguments{
\item{RELEVES}{Variables of the restoration sites data matrix}

\item{REF}{Variables of reference sites data matrix}

\item{DISTANCES}{Object resulting from the analysis of the \code{\link{ChosseRef::DissRef3}} function.}

\item{METHOD}{Dissimilarity index, partial match to "manhattan", "euclidean", "canberra", "clark", "bray", "kulczynski", "jaccard", "gower", "altGower", "morisita", "horn", "mountford", "raup", "binomial", "chao", "cao" or "mahalanobis". Cf \code{\link{vegan::vegdist}} function}

\item{COUL_RELEVES}{Color of RELEVES names}

\item{COUL_Rel_variable}{If TRUE: color change for each RELEVES}

\item{COUL_REF}{Color of REF names}

\item{COUL_Seg}{Color of links}

\item{COUL_Seg_variable}{If TRUE: colors change for each link}

\item{LINK_NUMBER}{Number of link drawn. If "N_REF" then for each RELEVES, the number of link drawn is the number given in \code{N_REF}. If "DIST_MIN"
only reference that have distance to restoration sites lower than the value given in \code{DIST_MIN} are drawn.}

\item{N_REF}{Number of REF to display a link between REF and RELEVES}

\item{DIST_MIN}{Minimum distance value to draw a link between REF and RELEVES}

\item{VAL_DIST}{Providing or not (TRUE by default) the distance value display}

\item{DECAL}{Distance between the printing of the distance and the link}
}
\description{
Display similarity link (previously calculated with \code{\link{ChosseRef::DissRef3}} function) between plots and a group of reference plot
}
\examples{
# ------------  Creating the data needed for the example --------------------
library(vegan)
data("dune") #downloading of dune data (cf vegan)
data("dune.env") #downloading of dune.env data (cf vegan)
# keeping only the numeric variables :
dune.env <- data.frame(A1 = dune.env$A1, 
                       Moisture =  as.numeric(as.vector(dune.env$Moisture)),
                       Manure = as.numeric(as.vector(dune.env$Manure)))
                                          
# Creating a vector indicating which plots are the potential references and which ones are the restored sites
sites <- factor(c(rep("Rest",5),rep("Ref",15)))
# Creating a vector with the plot names
sites_names <- paste(sites,c(1:5,1:15))

#Poviding names to rows (useful for the outputs)
row.names(dune.env) <- sites_names
row.names(dune) <- sites_names
dune.envRest <- dune.env[sites=="Rest",]
dune.envRef <- dune.env[sites=="Ref",]

# --------------  Calculating reference dissimilarities ---------------------
Distances <- DissRef3(RELEVES = dune.envRest, REF = dune.envRef, METHOD = "euclidean", DUPLICATES = FALSE)
Distances

# ----------------  Plotting reference dissimilarities ----------------------
Diss_Ref_Plot(RELEVES = dune.envRest, REF = dune.envRef, DISTANCES = Distances, LINK_NUMBER = "N_REF", N_REF = 3)
}
