% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptFonctionsSourcesVpack.r
\name{DissRef}
\alias{DissRef}
\title{Dissimilarite a une reference}
\usage{
DissRef(RELEVES, ETENDUEREL, ETENDUEREF, method, binary = T)
}
\arguments{
\item{RELEVES}{tableau de releves especes en colonnes, releves en lignes}

\item{ETENDUEREL}{etendues des releves sur lesquels le calcul sera fait (exemple 1:9 OU 3:11)}

\item{ETENDUEREF}{etendues des releves de reference (exemple 1:3)}

\item{method}{methode a utiliser cf l'aide de \code{\link{vegdist}}}

\item{binary}{T ou F selon que l'on travaille en presence absence ou non (T par defaut)}
}
\value{
Renvoie une liste de valeurs, une par releves chaque valeur correspond a la moyenne des indices qui ont ete calcules entre ce releves et tous les releves de references
}
\description{
Calcul de moyennes d'indices de similarite/dissimilarite entre des releves et plusieurs releves de references
}
\seealso{
\code{\link{vegdist}} \code{\link{raup.calc}}
}
