% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptFonctionsSourcesVpack.r
\name{ecart.ctrl}
\alias{ecart.ctrl}
\title{Ecart à un témoin par site}
\usage{
ecart.ctrl(VARIABLE, FACTEUR, CAT_COMP, SITE)
}
\arguments{
\item{VARIABLE}{vecteur ou data.frame, doit être numérique}

\item{FACTEUR}{facteur dont une des modalité est le témoin}

\item{CAT_COMP}{identifiant du témoin}

\item{SITE}{facteur donnant les identifiants du site}
}
\value{
un objet de la même forme que VARIABLE
}
\description{
calcule l'écart à un témoin par site pour une variable ou un groupe de variable
}
\examples{
var1 <- c(4,2,5,8,1,2,5,5)
var2 <- c(5,1,2,6,7,2,1,3)
resattendu1 <-c(0,-2,1,4,0,1,4,4)
resattendu2 <-c(0,-4,-3,1,0,-5,-6,-4)
facteur <- factor(rep(c("TEM","TRAI1","TRAI2","TRAI3"),2))
site <- factor(rep(c("site1","site2"),each=4))
tab <- data.frame(var1, var2, facteur, site)
ecart.ctrl(VARIABLE = tab$var1, FACTEUR = tab$facteur, CAT_COMP = "TEM", SITE = tab$site)
ecart.ctrl(VARIABLE = tab[,1:2], FACTEUR = tab$facteur, CAT_COMP = "TEM", SITE = tab$site)
}
