% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptFonctionsSourcesVpack.r
\name{Especes_dominantes}
\alias{Especes_dominantes}
\title{Affichage espèces dominantes}
\usage{
Especes_dominantes(RELEVES, SEL_RELEVES, NOMS_RELEVES, N)
}
\arguments{
\item{RELEVES}{le tableau de relevés}

\item{SEL_RELEVES}{les relevés selectionnés}

\item{NOMS_RELEVES}{le nom des relevés}

\item{N}{le nombre d'espèces dominantes à conserver}
}
\description{
Affichage espèces dominantes
}
\examples{
library(vegan)
data("dune")
TabSpDominantes <- Especes_dominantes(RELEVES = dune, SEL_RELEVES = c(4,7,8),
                                      NOMS_RELEVES = c(1:ncol(dune)), N = 10)
}
