% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptFonctionsSourcesVpack.r
\name{Especes_Typiques}
\alias{Especes_Typiques}
\title{Especes_Typiques}
\usage{
Especes_Typiques(
  RELEVES,
  FAC,
  MOD = "All",
  NOMBRE,
  NOMBRESTRICT = FALSE,
  METRIQUE = "Frequence"
)
}
\arguments{
\item{RELEVES}{le tableau de relevés}

\item{FAC}{vecteur du facteur donnant les catégories}

\item{MOD}{modalité du facteur FAC à afficher, si rien n'est mentionné, réalisé pour chaque modalité}

\item{NOMBRE}{nombre d'espèces les plus abondantes à afficher (possibilité d'avoir plus d'espèces affichées si ex aequo sauf si NOMBRESTRICT = TRUE)}

\item{NOMBRESTRICT}{affichage ou non des ex aequo}

\item{METRIQUE}{métrique utilisée pour le classement. Par défaut c'est la fréquence, i.e. le pourcentage de relevés dans laquelle l'espèce est présente. Sinon c'est l'abondance moyenne à l'échelle de l'ensemble des relevés de la modalité (pas uniquement ceux où l'espèce est présente).}
}
\description{
Especes_Typiques
}
\examples{
library(vegan)
data("dune")
Especes_Typiques(RELEVES = dune, FAC = dune.env$Use, NOMBRE = 5)
Hay <- Especes_Typiques(RELEVES = dune, FAC = dune.env$Use, NOMBRE = 3,
                 MOD = "Hayfield", NOMBRESTRICT = TRUE)
Hay
}
