% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptFonctionsSourcesVpack.r
\name{Heures}
\alias{Heures}
\title{Passer d'un format d'heure en décimale à du texte d'heure normales}
\usage{
Heures(heure_dec)
}
\arguments{
\item{heure_dec}{heures en décimales}
}
\value{
un texte avec l'heure en format horaire du type : "10h07"
}
\description{
Passer d'un format d'heure en décimale à du texte d'heure normales
}
\examples{
Heures(10.11)
Heures(29.9)
}
