% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptFonctionsSourcesVpack.r
\name{IC}
\alias{IC}
\title{Intervalle de confiance}
\usage{
IC(x)
}
\arguments{
\item{x}{un vecteur}
}
\value{
une valeur d'intervalle de confiance
}
\description{
Calcul d'un intervalle de confiance a 5 pour cent
}
\examples{
A<-rnorm(45,3,2)
IC(A)

}
