% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptFonctionsSourcesVpack.r
\name{label.corNMDS}
\alias{label.corNMDS}
\title{Affichage d'especes correlees}
\usage{
label.corNMDS(ANALYSE, RELEVES, p1 = 0.05, p2 = 0.05, coef = 1, add = F)
}
\arguments{
\item{ANALYSE}{Objet resultant d'une analyse multivariee \code{\link{dudi.pca}}, \code{\link{dudi.coa}}, \code{\link{metaMDS}} ou \code{\link{decorana}}}

\item{RELEVES}{les releves ayant servi a faire l'analyse multivariee}

\item{p1}{a p value minimale de correlation avec l'axe 1 pour etre affichee}

\item{p2}{la p value minimale de correlation avec l'axe 2 pour etre affichee}

\item{coef}{un coef multiplicateur (pour exploser ou non la dispersion dans le plan)}

\item{add}{par defaut, creer un graphe seul, si add=T, les noms d'especes sont ajoute au graphe deja existant}
}
\description{
Calcul les correlations des especes avec les axes puis n'affiche que les plus correlees
}
\details{
Attention, des erreurs souvent lorsqu'il y a des especes avec 0 occurences!!!! (hein Julie!)
}
\seealso{
\code{\link{label.cor}} apres  \code{\link{dudi.pca}} ou \code{\link{dudi.coa}} ;
\code{\link{label.corNMDS}} apres une NMDS avec  \code{\link{metaMDS}} ;
 \code{\link{label.corDCA}} apres une DCA avec  \code{\link{decorana}}
}
