% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptFonctionsSourcesVpack.r
\name{label.corV2}
\alias{label.corV2}
\title{Affichage d'espèces correlées}
\usage{
label.corV2(
  COO_ESP,
  COO_REL,
  RELEVES,
  METHOD = "defaut",
  P = "defaut",
  RHO = "defaut",
  N = "defaut",
  COEF = 1,
  ADD = F,
  ...
)
}
\arguments{
\item{COO_ESP}{Coordonnées des espèces suites à une ordination, pour les deux premiers axes, après dudi.pca et dudi.coa c'est ANALYSE$co[,1:2], après decorana c'est ANALYSE$cproj[,1:2], après metaMDS c'est ANALYSE$species[,1:2], etc.}

\item{COO_REL}{Coordonnées des relevés suites à une ordination, pour les deux premiers axes, après dudi.pca et dudi.coa c'est ANALYSE$li[,1:2], après decorana c'est ANALYSE$rproj[,1:2], après metaMDS c'est ANALYSE$points[,1:2], etc.}

\item{RELEVES}{les relevés ayant servi à faire l'analyse multivariée}

\item{METHOD}{Méthode de choix des espèces à afficher, soit "P" (défaut) - les espèces dont le p du test de spearman sont inférieures à la valeur de P  sur au moins un axe sont affichées, soit "RHO" - les espèces dont le Rho du test de spearman sont supérieur à RHO sur au moins un axe sont affichées, soit "N_base_P" - les N espèces dont le P sont les plus faibles sont affichées, soit "N_base_RHO" - les N espèces dont les RHO sont les plus fort sont affichées.}

\item{P}{la p value maximum pour être affichée}

\item{RHO}{le Rho minimum pour être affichée}

\item{N}{le nombre d'espèces à afficher, attention, si des ex-aequo, pour ne pas surcharger, il y a moins d'espèces affichées que le N choisi}

\item{COEF}{un coef multiplicateur (pour exploser ou non la dispersion dans le plan)}

\item{ADD}{par defaut, créer un graphe seul, si add=T, les noms d'especes sont ajouté au graphe déjà existant}

\item{...}{les autres arguments de la fonction text}
}
\description{
Calcul les corrélations des espèces avec les axes puis n'affiche que les plus corrélées
}
\examples{
# téléchargement des données de vegan
data("dune")
NMDS <- metaMDS(dune)
label.corV2(COO_ESP = NMDS$species, COO_REL = NMDS$points, RELEVES = dune, METHOD = "RHO", RHO = 0.5)
# En fonction du p, et sans graphe existant :
label.corV2(COO_ESP = NMDS$species[,1:2], COO_REL = NMDS$points[,1:2],
RELEVES = dune, METHOD = "P", P = 0.05)
# En fonction du rho et avec graphe existant :
plot(NMDS$points[,1:2])
label.corV2(COO_ESP = NMDS$species[,1:2], COO_REL = NMDS$points[,1:2],
            RELEVES = dune, METHOD = "RHO", RHO = 0.5, ADD = T)
Sans graphe précédent, par un nombre d'espèces en fonction du p :
label.corV2(COO_ESP = NMDS$species[,1:2], COO_REL = NMDS$points[,1:2],
            RELEVES = dune, METHOD = "N_base_P", N = 10)
# Sans graphe précédent, par un nombre d'espèces en fonction du rho :
label.corV2(COO_ESP = NMDS$species[,1:2], COO_REL = NMDS$points[,1:2],
RELEVES = dune, METHOD = "N_base_RHO", N = 15)
# Sans graphe précédent, par un nombre d'espèces en fonction du rho et en utilisant tout un tas d'autres arguments de la fonction `text`:
label.corV2(COO_ESP = NMDS$species[,1:2], COO_REL = NMDS$points[,1:2],
RELEVES = dune, METHOD = "N_base_RHO", N = 15, 
col="cadetblue", cex=1.5, font = 3)
# Et pour voir les effets de l'argument `COEF` :
plot(NMDS$species, ylim = c(-2, 2), xlim = c(-2,2))
label.corV2(COO_ESP = NMDS$species[,1:2], COO_REL = NMDS$points[,1:2],
            RELEVES = dune, METHOD = "N_base_RHO", N = 15, 
                        col="cadetblue", cex=0.5, COEF = 0.5, ADD = T)
label.corV2(COO_ESP = NMDS$species[,1:2], COO_REL = NMDS$points[,1:2],
            RELEVES = dune, METHOD = "N_base_RHO", N = 15, 
                        col="cadetblue2", cex=1, COEF = 1.5, ADD = T)
label.corV2(COO_ESP = NMDS$species[,1:2], COO_REL = NMDS$points[,1:2],
            RELEVES = dune, METHOD = "N_base_RHO", N = 15, 
                        col="cadetblue4", cex=1.5, COEF = 2, ADD = T)

}
