% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptFonctionsSourcesVpack.r
\name{modif_coul}
\alias{modif_coul}
\title{modif_coul}
\usage{
modif_coul(COULEUR, mods = 0.5, modv = 0.5, modh = 0.5, alpha = 1)
}
\arguments{
\item{COULEUR}{une couleur, soit en nom de couleur R ("cadetblue"), soit en code hexadécimal}

\item{mods}{coefficient modérateur de la saturation, sous 0.5 ça désature, au dessus, ça sature}

\item{modv}{coefficient modérateur de la brillance, sous 0.5 ça fonce, au dessus, ça éclaircit}

\item{modh}{coefficient modérateur de la teinte, ça tourne en rond..}

\item{alpha}{transparence, sous 1, ça devient transparent}
}
\value{
un code hexadécimal de couleur
}
\description{
modif_coul
}
