% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptFonctionsSourcesVpack.r
\name{MultiDyn}
\alias{MultiDyn}
\title{Tracer la dynamique en multivariee}
\usage{
MultiDyn(MultCOO, FACTime, FACT2, Axis = c(1, 2), LTY = 1, CEX = 1,
  COL = 1, ARR = F, LARR = 0.1, new = "yes", ...)
}
\arguments{
\item{MultCOO}{Les coordonnees de points sortant d'une analyse multivariee, par exemple si apres AFC sous ade4 (nommee \code{AFC_calc}) : \code{AFC_calc$li} par exemple si apres NMDS sous vegan (nommee \code{NMDS_calc}) : \code{NMDS$points}}

\item{FACTime}{le facteur temps, chaque modalit? corresponds un point de la ligne de trajectoire, il faut imperativement les avoir classe dans l'ordre}

\item{FACT2}{le facteur qui permet de regrouper des points, pour chaque modalite sera  representee une trajectoire, le nom de la modalite est affiche au debut de chaque trajectoire}

\item{Axis}{(facultatif, defaut: c(1,2)) precise les numero des axes a tracer (correspondant au numero des colonnes de \code{MultCOO})}

\item{LTY}{(facultatif, defaut: 1) le format des lignes, cf \code{lty}}

\item{CEX}{(facultatif, defaut: 1) la taille du texte, cf \code{cex}}

\item{COL}{(facultatif, defaut: 1) la couleur des lignes et du texte, si une seule couleur est precisee, toutes les trajectoires sont de la meme couleur, si plusieurs couleurs sont precisees, une couleur par trajectoire}

\item{ARR}{(facultatif, defaut: F) trace ou non une fleche a la fin de la trajectoire si \code{ARR=T}}

\item{LARR}{(facultatif, defaut: 0.1) si \code{ARR=T}, determine la taille de la fleche}

\item{new}{si new="yes", trace les fleches sur une nouvelle fenetre (defaut), sinon, sur le graphe existant}

\item{...}{d'autres arguments applicables a la fonction \code{\link{plot}}}
}
\description{
Trace les trajectoires de barycentres de modalites apres une analyse multivariee
}
