% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptFonctionsSourcesVpack.r
\name{multivar.domi}
\alias{multivar.domi}
\title{multivar.domi}
\usage{
multivar.domi(
  COO_REL,
  RELEVES,
  ESP,
  COL_NULL = "gray",
  COL_LOW = "blue",
  COL_HIGH = "red",
  cex = 1,
  ADD = FALSE,
  ...
)
}
\arguments{
\item{COO_REL}{Coordonnées des espèces suites à une ordination, pour les deux premiers axes, après dudi.pca et dudi.coa c'est ANALYSE$co[,1:2], après decorana c'est ANALYSE$cproj[,1:2], après metaMDS c'est ANALYSE$species[,1:2], etc.}

\item{RELEVES}{les relevés ayant servi à faire l'analyse multivariée}

\item{ESP}{l'espèce à afficher}

\item{COL_NULL}{couleur des points des relevés sans l'espèce}

\item{COL_LOW}{couleur pour les abondances faibles}

\item{COL_HIGH}{couleur pour les abondances fortes}

\item{cex}{taille des points de l'espèce}

\item{ADD}{par defaut, créer un graphe seul, si add=T, les noms d'especes sont ajouté au graphe déjà existant}

\item{...}{les autres arguments de la fonction plot (ne s'applique pas aux points de l'espèces, affichés par la fonction points..)}
}
\description{
multivar.domi
}
\examples{
data("dune")
ANALYSE_NMDS <- metaMDS(dune, trace = FALSE)
multivar.domi(COO_REL = ANALYSE_NMDS$points, RELEVES = dune, ESP = "Achimill")
}
