% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptFonctionsSourcesVpack.r
\name{multivar.polyg}
\alias{multivar.polyg}
\title{multivar.polyg}
\usage{
multivar.polyg(
  ANAcoo,
  FAC,
  pch = 1,
  col_dot = 1,
  col_fill = NA,
  col_text = 1,
  col_bord = 1,
  cex_lab = 1,
  font_lab = 1,
  dot = "yes",
  lab = "yes",
  new = "yes",
  sep = "no",
  ...
)
}
\arguments{
\item{ANAcoo}{coordonnees des points d'une analyse multivariees (__$points pour une NMDS, __$li pour une AFC, etc.)}

\item{FAC}{facteur qui permettra de separer les polygones}

\item{pch}{symbole des points (noir par defaut)}

\item{col_dot}{couleur des points (noir par defaut)}

\item{col_fill}{couleur du remplissage des polygone (transparent par defaut)}

\item{col_text}{couleur des lignes du texte (noir par defaut)}

\item{col_bord}{couleur des lignes du polygone (noir par defaut)}

\item{cex_lab}{taille des noms de modalites (si lab="yes")}

\item{font_lab}{police des étiquettes}

\item{dot}{si dot="yes", trace les points de chaque individu, "yes" par defaut}

\item{lab}{si lab="yes", affiche le nom de la modalite au niveau du barycentre de ses points, "yes" par defaut}

\item{new}{si new="yes", trace les polygones sur une nouvelle fenetre (defaut), sinon, sur le graphe existant}

\item{sep}{si sep="yes", trace chaque polygone sur une fenetre graphique differente, "no" par defaut}

\item{...}{possibilite de rajouter des arguments comme main,ylab,etc. associees a plot}
}
\description{
Dessine des polygones autour des points en fonction des modalites d'un facteur
}
\examples{
library(vegan)
data(varespec)
data(varechem)
facteur_Alu<-factor(ifelse(varechem$Al<50,"Al50",
                           ifelse(varechem$Al<110,"Al110",
                                  ifelse(varechem$Al<230,"Al230","Al++"))))
ord <- metaMDS(varespec)
multivar.polyg(ord$points,facteur_Alu)
}
