% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptFonctionsSourcesVpack.r
\name{occ}
\alias{occ}
\title{Graphe d'occ}
\usage{
occ(VAR, CEX = 0.5, postext = 1, ec = 5, reor = T, x11 = F, ...)
}
\arguments{
\item{VAR}{le vecteur avec les mots}

\item{CEX}{expansion factor des noms des occurences}

\item{postext}{la position du nombre de (cf axe des y)}

\item{ec}{ecart avec la marge \code{\link{par}} et son parametre \code{oma}}

\item{reor}{si \code{T}, dans l'ordre croissant, sinon, dans l'ordre donne}

\item{x11}{si \code{T}, une nouvelle fenetre est ouverte, sinon, non}

\item{...}{d'autres arguments de la fonction \code{\link{barplot}}}
}
\description{
Fonction specifique dans le cadre de l'analyse des resumes de REVER... permet de faire
un graphe d'occurence des mots en donnant le vecteur
}
\examples{
Texte_exemple=c("Banana","Fraise","Chocolat","Chocolat","Chocolat","Fraise","Banana","Banana","Banana","Banana")
occ(Texte_exemple,main="Exemple parfum de glaces...")
}
