% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptFonctionsSourcesVpack.r
\name{point.plot}
\alias{point.plot}
\title{point.plot}
\usage{
point.plot(variable, Facteur, lettres = c(""), type = "p", ecart = "sem",
  ylim = "NP", xlim = "NP", add = "FALSE", ...)
}
\arguments{
\item{variable}{un vecteur avec les valeurs des variables}

\item{Facteur}{un facteur donnant les differentes modalites}

\item{lettres}{(facultatif) liste de type c("text","text") qui figurera au dessus des barres}

\item{type}{Type de liaison entre les points ("b": points reliés - esp points-ligne;"o": points reliés - pas d'espace; "l": que les lignes;"p" : que les points;"s": en marches}

\item{ecart}{(facultatif) type de calcul de la barre d'erreur (sem, IC, sd, var, etc.)}

\item{ylim}{comme pour la fonction plot(), permet de preciser les limites de l'axe des ordonnees}

\item{...}{possibilite de rajouter des arguments comme col,main,ylab,etc. associees a barplot}
}
\description{
Fait un diagramme en points avec la moyenne et les barres d'erreurs
}
\details{
Co-écrite avec Julie Chenot!
}
