% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptFonctionsSourcesVpack.r
\name{radarchart2}
\alias{radarchart2}
\title{Diagramme radar avec possibilite de ponderer...}
\usage{
radarchart2(
  df,
  axistype = 0,
  seg = 4,
  pty = 16,
  pcol = 1:8,
  plty = 1:6,
  plwd = 1,
  pdensity = NULL,
  pangle = 45,
  pfcol = NA,
  cglty = 3,
  cglwd = 1,
  cglcol = "navy",
  axislabcol = "blue",
  title = "",
  maxmin = TRUE,
  na.itp = TRUE,
  centerzero = FALSE,
  vlabels = NULL,
  vlcex = NULL,
  caxislabels = NULL,
  calcex = NULL,
  paxislabels = NULL,
  palcex = NULL,
  coef = 1,
  ...
)
}
\description{
Diagramme radar avec possibilite de ponderer...
}
\details{
Fonction en cours de production... pas d'aide pour le moment... Modif de la fonction radarchart du package (fmsb) avec ajout de l'argument coef  pour ponderer les variables...

Minato Nakazawa (2014). fmsb: Functions for medical statistics book with some demographic data. R package version 0.5.1.

http://CRAN.R-project.org/package=fmsb

neccessite de charger la library fmsb
}
