% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptFonctionsSourcesVpack.r
\name{rar.rm}
\alias{rar.rm}
\title{Suppression d'especes rares}
\usage{
rar.rm(tableau.AD, n)
}
\arguments{
\item{tableau.AD}{un tableau de contingence}

\item{n}{un nombre minimal d'occurence}
}
\value{
le meme taleau de contingence, mais sans les especes dont le nombre d'occurence est inferieur a n
}
\description{
Supprime les especes n'ayant pas un certain nombre d'occurences
}
\examples{
#a construire...

}
