% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptFonctionsSourcesVpack.r
\name{raup.calc}
\alias{raup.calc}
\title{valeur de similarite/dissimilarite}
\usage{
raup.calc(tableau, B, A, method, binary = T)
}
\arguments{
\item{tableau}{tableau de releves especes en colonnes, releves en lignes}

\item{B}{numero du releves dont le numero de ligne est le plus eleve}

\item{A}{numero du releves dont le numero de ligne est le plus faible}

\item{method}{ethode a utiliser cf l'aide de \code{\link{vegdist}}}

\item{binary}{T ou F selon que l'on travaille en presence absence ou non (T par defaut)}
}
\value{
une valeur de similarite/dissimilarite
}
\description{
calcule un indice de similarite/dissimilarite entre 2 releves
}
\seealso{
\code{\link{vegdist}}
}
