% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptFonctionsSourcesVpack.r
\name{Remplacer}
\alias{Remplacer}
\title{Permet, pour chaque valeur possible d'un vecteur, de les remplacer par des valeurs predefinies}
\usage{
Remplacer(LISTE, VALEURS, VALEURS_NEW)
}
\arguments{
\item{LISTE}{un vecteur}

\item{VALEURS}{les valeurs possibles de LISTE (possibilite d'utiliser unique(LISTE) ou levels(LISTE))}

\item{VALEURS_NEW}{les valeurs de remplacements, a donner dans le meme ordre que VALEURS}
}
\value{
un vecteur ayant la meme longueur que LISTE
}
\description{
Permet, pour chaque valeur possible d'un vecteur, de les remplacer par des valeurs predefinies
}
\examples{
chiffres<-c(1,1,1,2,4,3,2,1,2,3,3,1,3,2)
valeurs_avant<-c(1,2,3,4)
valeurs_apres<-c("A","B","C","D")

Remplacer(chiffres,valeurs_avant,valeurs_apres)
}
