% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptFonctionsSourcesVpack.r
\name{RepNAbyMean}
\alias{RepNAbyMean}
\title{Replace NA by mean}
\usage{
RepNAbyMean(TAB)
}
\arguments{
\item{TAB}{a table with numeric values with possible NAs}
}
\value{
the same table as TAB but without anymore NA values
}
\description{
replace NA value by the mean of the column
}
\examples{
#Le tableau pour tester :
var1<-c(8,4,5,6,NA)
var2<-c(1,NA,2,1,3)
var3<-c(NA,11,10,NA,12)
tab_essai<-data.frame(var1,var2,var3)

#L'utilisation de la fonction :
tab_essai_new<-RepNAbyMean(tab_essai)
}
