% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptFonctionsSourcesVpack.r
\name{sem}
\alias{sem}
\title{sem}
\usage{
sem(x)
}
\arguments{
\item{x}{un vecteur}
}
\value{
une valeur d'erreur standard
}
\description{
Calcul de l'erreur standard
}
\details{
formule ecart-type/racine du nombre de valeur. Attention, il faudrait rajouter un argument pour utiliser la formule n-1 quand neccessaire...
}
\examples{
A<-rnorm(45,3,2)
sem(A)

}
