% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptFonctionsSourcesVpack.r
\name{Sim_Clust_Ordre}
\alias{Sim_Clust_Ordre}
\title{Similarite releves-syntaxon (cf these Julie)}
\usage{
Sim_Clust_Ordre(TAB_FREQ_CLU, NOM_CLU, TAB_ORDRE, METH = "bray", BINA = F)
}
\arguments{
\item{TAB_FREQ_CLU}{tableau de releves des especes en colonnes et des frequences des especes de chaque cluster en ligne (2 colonnes de variables et colonnes especes). Les valeurs correspondent a la frequence des especes pour chaque cluster}

\item{NOM_CLU}{Nom du cluster choisit}

\item{TAB_ORDRE}{tableau des releves de references qui definit les differentes alliances vegetales (les especes en colonne et les alliances en ligne) (4 colonnes de variables et des colonnes especes). Les valeurs correspondent a la frequence des especes pour chaque alliance.}

\item{METH}{methode de similarite a utiliser (par defaut = bray)}

\item{BINA}{binarite T ou F selon que l'on travaille en presence/absence ou non (par default = FALSE)}
}
\value{
renvoie un tableau avec  une valeur de similarite (1-indice de Bray Curtis) et un rang (l'alliance la plus similaire) pour chaque alliance en comparaison avec le cluster choisit
}
\description{
Calcul la similiarite des releves entre un tableau de releves (releves regroupes dans un cluster) et un groupement vegetal de reference (de type alliance vegetale)
}
\examples{
# Exemple d'application
# chargement du package vegan
library(vegan)
# Creation de tableaux de variables / facteurs
tord<-data.frame(list(nom_ordre = c("Sisymbrietea", "Sisymbrietea", "Sisymbrietea"),
                      Ordres = c("O.1.", "O.1.", "O.1."),
                      alliances = c("A.1.2.", "A.1.3.", "A.1.4."),
                      Nb_syntaxons = c(22,28,32),
                      Sp1=c(1,0,0),
                      Sp2=c(1,1,0),
                      Sp3=c(0,1,1),
                      Sp4=c(0,1,0)))
tord
tclu<-data.frame(list(Num_releve = c("3", "4", "6"),All_rel = c("Carriere", "Carriere", "Carriere"),
                      Sp1=c(0,0,0.5),
                      Sp2=c(1,0,1),
                      Sp3=c(1,0,0.1),
                      Sp4=c(0.5,0,0),
                      Sp5=c(0,1,0),
                      Sp6=c(0,1,0)))
tclu
# Utilisation de la fonction pour le cluster 6 (=Num_releve) par exemple
Sim_Clust_Ordre(tclu,NOM_CLU="6",tord,METH="bray",BINA=F)
}
