% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptFonctionsSourcesVpack.r
\name{structure.plot}
\alias{structure.plot}
\title{structure community plots}
\usage{
structure.plot(
  INDICE,
  FACTOR,
  MULTI = T,
  MTITLE = "",
  ABMAX = 5,
  col1 = "grey60",
  col2 = "white",
  col3 = "red",
  col4 = "orange",
  noms = "T",
  cex_noms = 1,
  ...
)
}
\arguments{
\item{INDICE}{An object issued from \code{\link{ComStructIndices}} function}

\item{FACTOR}{A factor list, a barplot of species mean abundances will be performed for each factor level. If no factor is specified, MULTI=F should be specified.}

\item{MULTI}{If no factor is specified, MULTI=F should be specified}

\item{MTITLE}{Main title of the plot}

\item{ABMAX}{Numerical value of the maximum abundance}

\item{col1}{Colour information for the Reference mean abundances barplot}

\item{col2}{Colour information for the Reference mean abundances in assessed community barplot, i.e. "missing abundances"}

\item{col3}{Colour information for the abundances of target species in the assessed community}

\item{col4}{Colour information for the "higher abundances" in the assessed community}

\item{noms}{If other than "T", species names are not given}

\item{cex_noms}{expansion factor for species names}

\item{...}{other parameters from the \code{\link{barplot}} function}
}
\description{
Performs a barplot of abundances of species in assessed community compared to a REFerence community
}
\seealso{
\code{\link{ComStructIndices}}
}
