% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptFonctionsSourcesVpack.r
\name{Tableau_recap}
\alias{Tableau_recap}
\title{Tableau recapitulatif de donnees}
\usage{
Tableau_recap(VAR, FAC, ROUND, LETTRES = c(""), ERROR = sem)
}
\arguments{
\item{VAR}{le vecteur avec les mots}

\item{FAC}{un facteur donnant les differentes modalites}

\item{ROUND}{un nombre de decimal pour l'arrondi des moyennes et erreurs}

\item{LETTRES}{(facultatif) liste de type c("text","text") qui figurera au dessus des barres}

\item{ERROR}{(facultatif) type de calcul de la barre d'erreur (sem, IC, sd, var, etc.)}
}
\description{
Permet de creer automatiquement un tableau de moyenne, erreur
}
\examples{
#Donnees
biomasse<-c(rnorm(35,20,5),rnorm(35,15,3),rnorm(35,5,3))
traitement<-factor(rep(c("T1","T2","T3"),each=35))
Tableau_recap(biomasse,traitement,2)
#Si on a fait un test post-hoc :
anov<-aov(biomasse~traitement)
shapiro.test(anov$residuals)
summary(anov)
TukeyHSD(anov)
#les lettres sont donc a, b et c :
Tableau_recap(biomasse,traitement,2,c("a","b","c"))
}
