% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptFonctionsSourcesVpack.r
\name{Time.factor.plot}
\alias{Time.factor.plot}
\title{Time.factor.plot}
\usage{
Time.factor.plot(
  TIME,
  FACTOR,
  variable,
  xlab = "TIME",
  ylab = "variable",
  pch = NULL,
  couleur = NULL,
  lty = NULL,
  lwd = NULL,
  xlim = NULL,
  ylim = NULL,
  Posit_leg = "topleft",
  linktyp = "b",
  etoiles = c(""),
  ...
)
}
\arguments{
\item{TIME}{un vecteur comportant les donnees "temps" ou toute autre variable quantitative qui sera tracee en abscisse}

\item{FACTOR}{un vecteur comportant les donnees du facteur (il sera transformee en facteur qualitatif)}

\item{variable}{un vecteur comportant les donnees a tracer}

\item{xlab}{etiquette des abcsisses, par defaut: TIME}

\item{ylab}{tiquette des ordonnees, par defaut: variable}

\item{pch}{comme pour la fonction \code{\link{plot}}, permet d'assigner un symbole par modalite du facteur FACTOR}

\item{couleur}{permet d'assigner une couleur a chaque modalite du facteur FACTOR}

\item{lty}{permet d'assigner une forme de ligne par modalite du facteur FACTOR}

\item{lwd}{permet d'assigner une épaisseur de ligne par modalite du facteur FACTOR}

\item{xlim}{comme pour la fonction plot(), permet de preciser les limites de l'axe des abscisses}

\item{ylim}{comme pour la fonction plot(), permet de preciser les limites de l'axe des ordonnees}

\item{Posit_leg}{permet de preciser la position de la legende, soit c(x,y), soit "topleft", "bottomright", etc.}

\item{linktyp}{permet de pr?ciser quel type de lien entre les points... : \code{"p"} for points,\code{"l"} for lines, \code{"o"} for overplotted points and lines, \code{"b"}, \code{"c"})}

\item{etoiles}{permet de mettre du texte (ou des etoiles) au dessus de chacune des "TIME"}

\item{...}{autres arguments associes a la fonction \code{\link{plot}}}
}
\description{
Trace une courbe par modalite d'un facteur d'une variable en fonction du temps (moyenne + barres d'erreurs)
}
\examples{
# Exemple d'application :
# Creation des variables / facteurs
tim<-rep(c(1:4),each=6)
fac<-factor(rep(rep(c("A","B"),each=3),4))
vari<-c(rnorm(3,5,2),rnorm(3,3,2),rnorm(3,3,2),rnorm(3,8,2),
        rnorm(3,8,2),rnorm(3,10,2),rnorm(3,16,2),rnorm(3,12,2))

#Utilisation de la fonction :
Time.factor.plot(tim,fac,vari,etoiles=c("*","***","","***"))

}
