% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptFonctionsSourcesVpack.r
\name{better_arrows}
\alias{better_arrows}
\title{Trace des fleches suite a une ACP}
\usage{
better_arrows(ANALYSE, coef = 1, length = 0.1, angle = 15, col_text = 1,
  ...)
}
\arguments{
\item{ANALYSE}{objets issues d'una ACP avec dudi.pca (ade4)}

\item{coef}{un coef multiplicateur (pour exploser ou non la dispersion dans le plan)}

\item{length}{longueur de l'encoche des fleches, 0.1 = 10pourcent par defaut}

\item{angle}{angle des encoches des fleches, 15 = 15deg par defaut}

\item{col_text}{couleur du texte}

\item{...}{n'importe quel argument de la fonction arrows}
}
\description{
Trace des fleches suite a une ACP
}
\examples{
# Creation donnees :
sol<-c(rnorm(10,5,2),rnorm(10,15,3))
veg<-c(rnorm(20,21,5))
amphi<-c(rnorm(5,5,2),rnorm(5,5,3),rnorm(5,5,2),rnorm(5,5,3))
tablo<-data.frame(sol,veg,amphi)
# Faire l'ACP
library(ade4)
ACP<-dudi.pca(tab,scannf = F, nf = 2)
plot(ACP$li)
better_arrows(ACP,coef=2)
}
