% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptFonctionsSourcesVpack.r
\name{modif_coul}
\alias{modif_coul}
\title{modif_coul}
\usage{
modif_coul(COULEUR, mods = 0.5, modv = 0.5, modh = 0.5, alpha = 1)
}
\arguments{
\item{COULEUR}{une couleur, soit en nom de couleur R ("cadetblue"), soit en code hexadécimal}

\item{mods}{coefficient modérateur de la saturation, sous 0.5 ça désature, au dessus, ça sature}

\item{modv}{coefficient modérateur de la brillance, sous 0.5 ça fonce, au dessus, ça éclaircit}

\item{modh}{coefficient modérateur de la teinte, ça tourne en rond..}

\item{alpha}{transparence, sous 1, ça devient transparent}
}
\value{
un code hexadécimal de couleur
}
\description{
Permet de modifier des couleurs
}
\examples{
#Sans modification de couleur :
hist(rnorm(30),col="cadetblue")
#En modifiant la saturation :
hist(rnorm(30),col=modif_coul("cadetblue",mods=0.1))
hist(rnorm(30),col=modif_coul("cadetblue",mods=0.9))
#En modifiant la clarté :
hist(rnorm(30),col=modif_coul("cadetblue",modv=0.1))
hist(rnorm(30),col=modif_coul("cadetblue",modv=0.9))
#En modifiant la transparence :
hist(rnorm(30),col=modif_coul("cadetblue",alpha=0.1))
hist(rnorm(30),col=modif_coul("cadetblue",alpha=0.9))

}
