% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptFonctionsSourcesVpack.r
\name{point.plot}
\alias{point.plot}
\title{point.plot}
\usage{
point.plot(variable, Facteur, lettres = c(""), type = "p", ecart = "sem",
  ylim = "NP", xlim = "NP", add = "FALSE", ...)
}
\arguments{
\item{variable}{un vecteur avec les valeurs des variables}

\item{Facteur}{un facteur donnant les differentes modalites}

\item{lettres}{(facultatif) liste de type c("text","text") qui figurera au dessus des barres}

\item{type}{Type de liaison entre les points ("b": points reliés - esp points-ligne;"o": points reliés - pas d'espace; "l": que les lignes;"p" : que les points;"s": en marches}

\item{ecart}{(facultatif) type de calcul de la barre d'erreur (sem, IC, sd, var, etc.)}

\item{ylim}{comme pour la fonction plot(), permet de preciser les limites de l'axe des ordonnees}

\item{xlim}{comme pour la fonction plot(), permet de preciser les limites de l'axe des abscisses}

\item{add}{si jamais on veut que ça soit rajouté sur un graphique existant, il faut mettre "TRUE"}

\item{...}{possibilite de rajouter des arguments comme col,main,ylab,etc. associees a barplot}
}
\description{
Fait un diagramme en points avec la moyenne et les barres d'erreurs
}
\details{
Co-écrite avec Julie Chenot!
}
\examples{
#Donnees
biomasse<-c(rnorm(35,20,5),rnorm(35,15,3),rnorm(35,5,3))
traitement<-factor(rep(c("T1","T2","T3"),each=35))
#Graphique de base
point.plot(variable = biomasse,Facteur = traitement)
#En changeant les barres d'erreurs
point.plot(biomasse,traitement,ecart=sd)
#En ajoutant les lettres de post-hoc
point.plot(biomasse,traitement,lettres=c("a","ab","b"))
#En changeant des parametres classiques de barplot...
point.plot(biomasse,traitement,lettres=c("a","ab","b"),
           col=c("cadetblue"),
           xlab="Traitements herbicides",ylab="Biomasse",font.lab=3,
           main="Efficacite des traitement",pch=16,type="b",lty=3,lwd=4,cex=3)
#Si on veut ajouter à un graphe existant..
point.plot(biomasse/2,traitement,lettres=c("a","ab","b"),
           col=c("darkolivegreen"),pch=16,add="TRUE")

}
