% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptFonctionsSourcesVpack.r
\name{wheelscores}
\alias{wheelscores}
\title{Score wheel}
\usage{
wheelscores(
  lowercat,
  supercat,
  scores,
  col.score = "#a1d99b",
  col.null = "white",
  col.lines = "black",
  col.border = "grey",
  cex.sup = 1,
  col.super = NULL,
  cex.low = 0.7,
  circles = TRUE,
  ...
)
}
\arguments{
\item{lowercat}{A vector of scores labels}

\item{supercat}{A vector of categories labels}

\item{scores}{A numeric vector of scores (between 0 and 5)}

\item{col.score}{The color of achieved scores}

\item{col.null}{The color of unachieved scores}

\item{col.lines}{The color of lines and circles}

\item{col.border}{The color of the border containing categories labels}

\item{cex.sup}{An expansion factor for categories labels}

\item{col.super}{A vector of colors for each categories}

\item{cex.low}{An expansion factor for scores labels}

\item{circles}{(default = TRUE) if FALSE, doesn't draw intermediate circles}
}
\description{
Draw a score wheel inspired from the recovery wheel of SER Standards (McDonald et al., 2016)
}
\details{
uses function from plotrix package Lemon, J. (2006) Plotrix: a package in the red light district of R. R-News, 6(4): 8-12.
}
